Got the cleaner parse to mostly work.
There’s a very strange error when dealing with Enum.valueOf(). What’s
happening is that the signature has a whole other clause in front of
the arguments for some reason which I think defines the type that T
must be. I don’t really know how to handle that, so I’m just ignoring
it in the parsing.